properly deleting athletes from all joins

+++ XCoach/XCoach/Controllers/AthletesController.cs
@@ -151,9 +151,15 @@ namespace XCoach.Controllers
             {
                 return NotFound();
             }
+            var currentUser = await GetCurrentUserAsync();
 
             var athlete = await _context.Athletes
                 .Include(a => a.User)
+                .Include(a => a.AthleteRaces)
+                .ThenInclude(ar => ar.Athlete)
+                .Include(a => a.AthleteWorkouts)
+                .ThenInclude(aw => aw.Athlete)
+                .Where(a => a.Id == id)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (athlete == null)
             {
@@ -168,8 +174,32 @@ namespace XCoach.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var user = await GetCurrentUserAsync();
+            var userId = user.Id;
             var athlete = await _context.Athletes.FindAsync(id);
-            _context.Athletes.Remove(athlete);
+            var athleteWorkouts = _context.AthleteWorkouts;
+            var athleteRaces = _context.AthleteRaces;
+
+            foreach(AthleteWorkout wo in athleteWorkouts)
+            {
+                if(wo.AthleteId == athlete.Id && userId == athlete.UserId)
+                {
+                    athleteWorkouts.Remove(wo);
+                }
+                
+            }
+            foreach(AthleteRace ra in athleteRaces)
+            {
+                if(ra.AthleteId == athlete.Id && userId == athlete.UserId)
+                {
+                    athleteRaces.Remove(ra);
+                }
+            }
+            if(userId == athlete.UserId)
+            {
+                _context.Athletes.Remove(athlete);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
